Update Main Program
This has my menus set up as Dictionaries. I can then pull keys from the dictonary and I get to store all the object information which is super handy. I haven't added it yet, but I can add all the selections to a list, call it bill and use that to tally a bill and calculate tip easily.